fix: fix type casting for button position calculations.
Explicitly cast screen dimensions to float to ensure consistent positioning of the play button on all devices. This resolves potential issues caused by integer division in certain environments.

+++ src/main/java/org/example/telasDoJogo/TelaDeInicio.java
@@ -40,7 +40,7 @@ public class TelaDeInicio implements Screen {
         ImageButton buttonJogar = new ImageButton(new TextureRegionDrawable(new TextureRegion(buttonTexture)));
 
         buttonJogar.setSize(buttonJogar.getWidth()*0.5f, buttonJogar.getHeight()*0.2f);
-        buttonJogar.setPosition(Gdx.graphics.getWidth()/2-buttonJogar.getWidth()/2, Gdx.graphics.getHeight()/2-buttonJogar.getHeight()/2);
+        buttonJogar.setPosition((float) Gdx.graphics.getWidth() /2-buttonJogar.getWidth()/2, (float) Gdx.graphics.getHeight() /2-buttonJogar.getHeight()/2);
 
         buttonJogar.addListener(new ClickListener(){
